restructured files and folders; added run-all
