+++ src/ex/ex-readfile.js
@@ -1,14 +1,14 @@
 const Task = require('data.task')
 const futurize = require('futurize').futurize(Task)
 const { List } = require('immutable-ext')
-const execTask = require('./exec')
+const execTask = require('../helpers/exec')
 const fs = require('fs')
 const path = require('path')
 
 const readFile = futurize(fs.readFile)
 const files = [
-  path.join(__dirname, '/exec.js'),
-  path.join(__dirname, '/task.js')
+  path.join(__dirname, '/../helpers/exec.js'),
+  path.join(__dirname, '/../helpers/logger.js')
 ]
 const list = List(files)
