+++ src/ex/ex-reduce-regular.js
