+++ src/run-all.js
@@ -1,0 +1,8 @@
+require('./ex/ex-list')
+require('./ex/ex-readfile')
+require('./ex/ex-reduce-regular')
+require('./ex/ex-task')
+
+require('./async/ex-map-reduce-async')
+require('./async/map-reduce-async')
+require('./map-reduce-seq')
